determineNextPresetKeyをpreset.tsに移した
